[Generator] `start_index` & `seed_phrase` options
This change allows the user to specify the starting index at which
the keys should begin generating from, as well as the ability to
reuse a previous seed phrase.

This change allows me reuse a previously generated mnemonic phrase to
generate new validator keys starting at an arbitrary index of the HD
wallet.

```python
$ python eth2deposit/deposit.py \
    --num_validators=20 \
    --start_index=40 \
    --mnemonic_language=english \
    --chain=medalla \
    --seed_phrase="<REDACTED>" \
    --password="<REDACTED>"
```

Not a bug, but one caveate is that the user needs to be mindful of what
`start_index` value was used last. Thankfully, the validator index is
already included as part of the _keystore_ `json` file.